库存管理优化,库存低于预警值时，通知管理员及时补货

+++ sz-boot-parent/sz-service/sz-service-admin/src/main/java/com/sz/admin/restaurant/pojo/dto/InventoryUpdateDTO.java
@@ -1,5 +1,6 @@
 package com.sz.admin.restaurant.pojo.dto;
 
+import cn.idev.excel.annotation.ExcelProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
@@ -22,10 +23,10 @@ public class InventoryUpdateDTO {
     private String materialName;
 
     @Schema(description =  "当前容量")
-    private Integer currentStock;
+    private Double currentStock;
 
     @Schema(description =  "最小容量")
-    private Integer minStock;
+    private Double minStock;
 
     @Schema(description =  "计量单位")
     private String unit;
