+++ sz-boot-parent/sz-service/sz-service-admin/src/main/java/com/sz/admin/restaurant/pojo/po/DishRecipe.java
@@ -36,7 +36,7 @@ public class DishRecipe implements Serializable {
     private Long materialId;
 
     @Schema(description = "制作一份该菜品所需原材料的数量")
-    private BigDecimal materialQuantity;
+    private Double materialQuantity;
 
 
 }
